requirements.txt
openai>=1.71.0
gitpython>=3.1.44
python-dotenv>=1.1.0
aiohttp>=3.9.0
uvicorn>=0.24.0
anthropic>=0.7.0
anthropic-tools>=0.1.0
anthropic-bedrock>=0.1.0